+++ src/main/java/co/edu/usergioarboleda/alquiler/cabin/app/services/ScoreService.java
@@ -1,6 +1,7 @@
 package co.edu.usergioarboleda.alquiler.cabin.app.services;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,7 +19,7 @@ public class ScoreService {
         return repository.findAll();
     }
 
-    public Score getById(Integer id) {
+    public Optional<Score> getById(Integer id) {
         return repository.findById(id);
     }
 
@@ -29,6 +30,7 @@ public class ScoreService {
             if (repository.findById(score.getId()) == null) {
                 return repository.save(score);
             } else {
+                // throw new RuntimeException("Score not found");
                 return score;
             }
         }
@@ -36,28 +38,45 @@ public class ScoreService {
 
     public Score update(Score score) {
         if (score.getId() != null) {
-            Score newScore = repository.findById(score.getId());
-            if (newScore != null) {
+            Optional<Score> optionalScore = repository.findById(score.getId());
+            if (optionalScore.isPresent()) {
                 if (score.getScore() != null) {
-                    newScore.setScore(score.getScore());
+                    optionalScore.get().setScore(score.getScore());
                 }
                 if (score.getMessage() != null) {
-                    newScore.setMessage(score.getMessage());
+                    optionalScore.get().setMessage(score.getMessage());
                 }
-                return repository.save(newScore);
+                return repository.save(optionalScore.get());
             } else {
+                // throw new RuntimeException("Score not found");
                 return score;
             }
         } else {
+            // throw new RuntimeException("Score not found");
             return score;
         }
     }
 
-    public void delete(Score score) {
-        repository.delete(score);
+    public void deleteAll(Score score) {
+        if (score.getId() != null) {
+            Optional<Score> optionalScore = repository.findById(score.getId());
+            if (optionalScore.isPresent()) {
+                repository.delete(score);
+            } else {
+                // throw new RuntimeException("Score not found");
+            }
+        } else {
+            // throw new RuntimeException("Score not found");
+        }
+
     }
 
     public void deleteById(Integer id) {
-        repository.deleteById(id);
+        Optional<Score> optionalScore = repository.findById(id);
+        if (optionalScore.isPresent()) {
+            repository.delete(optionalScore.get());
+        } else {
+            // throw new RuntimeException("Score not found");
+        }
     }
 }
